backend/requirements.txt
asgiref~=3.8.1
Django~=5.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework~=3.15.2
djangorestframework_simplejwt==5.5.1
Markdown==3.8.2
pillow~=10.4.0
pip~=25.1
PyJWT==2.10.1
setuptools~=80.3.0
sqlparse~=0.5.1
typing_extensions~=4.13.2
tzdata==2025.2
django-smart-selects==1.7.2
pycountry~=24.6.1
utils~=1.0.2
numpy~=1.26.4
filelock~=3.18.0
colorama~=0.4.6
PyYAML~=6.0.1
Jinja2~=3.1.4
requests~=2.31.0
packaging~=24.2
pyparsing~=3.2.1
pytz~=2025.2